server-monitoring2/requirements.txt
Django
gunicorn
pyzmq